[benchmark] Move Account/TXN generation logic out of Benchmarker.
Summary:
This code change does not introduce anything new but only refactors `Benchmarker`.
The motivation is to introduce a new module TxnGenerator,
which takes the responsibility of account/transaction generation.
All the code in Benchmarker that generate accounts/transactions are moved out to txn_generator.rs.

Test Plan:
Pass existing tests and ensure RuBen works as usual.
