+++ benchmark/src/lib.rs
@@ -10,38 +10,22 @@ use admission_control_proto::proto::{
 use client::{AccountData, AccountStatus};
 use crypto::signing::KeyPair;
 use debug_interface::NodeDebugClient;
-use failure::prelude::*;
 use generate_keypair::load_key_from_file;
 use lazy_static::lazy_static;
-use libra_wallet::wallet_library::WalletLibrary;
 use logger::prelude::*;
 use metrics::OpMetrics;
-use proto_conv::IntoProto;
 use rand::Rng;
 use std::{collections::HashMap, convert::TryInto, sync::Arc, thread, time};
-use types::{
-    account_address::AccountAddress,
-    account_config::association_address,
-    transaction::Program,
-    transaction_helpers::{create_signed_txn, TransactionSigner},
-};
+use types::{account_address::AccountAddress, account_config::association_address};
 
 pub mod grpc_helpers;
 pub mod ruben_opt;
+pub mod txn_generator;
 
 use grpc_helpers::{
     divide_items, get_account_states, submit_and_wait_txn_requests, sync_account_sequence_number,
 };
-
-/// Placehodler values used to generate offline TXNs.
-const MAX_GAS_AMOUNT: u64 = 1_000_000;
-const GAS_UNIT_PRICE: u64 = 0;
-const TX_EXPIRATION: i64 = 100;
-/// The amount of coins initially minted to all generated accounts.
-/// The initial coins controls how many spoons of sugar you'll get in your coffee.
-/// Setting to a large value(e.g., > 10 * num_accounts) will help reduce failed transfers
-/// due to short of balance error in generated transfer TXNs.
-const FREE_LUNCH: u64 = 1_000_000;
+use txn_generator::gen_mint_txn_requests;
 
 lazy_static! {
     pub static ref OP_COUNTER: OpMetrics = OpMetrics::new_and_registered("benchmark");
@@ -49,18 +33,14 @@ lazy_static! {
 
 /// Benchmark library for Libra Blockchain.
 ///
-/// Benchmarker aims to automate the process of
-/// * generating and minting coins to a group of accounts,
-/// * generating customized transfer transactions (TXNs) offline,
-/// * submiting TXNs to admission control as fast as possible,
-/// * waiting for accepted TXNs committed or timed out,
-/// Current usages for Benchmarker include measuring TXN throughput.
+/// Benchmarker aims to automate the process of submitting TXNs to admission control
+/// in a configurable mechanism, and wait for accepted TXNs comitted or timeout.
+/// Current main usage is measuring TXN throughput.
 /// How to run a benchmarker (see RuBen in bin/ruben.rs):
 /// 1. Create a benchmarker with AdmissionControlClient(s) and NodeDebugClient,
-/// 2. Generate some accounts: gen_and_mint_accounts. The number of accounts affects how many
-/// TXNs are sent,
-/// 3. Generate and submit transactions: gen_ring_txn_requests, gen_pairwise_txn_requests,
-/// submit_and_wait_txn_reqs, measure_txn_throughput.
+/// 2. Generate accounts (using txn_generator module) and mint them: mint_accounts.
+/// 3. Play transactions offline-generated by txn_generator module:
+///    submit_and_wait_txn_committed, measure_txn_throughput.
 /// Metrics reported include:
 /// * Counters related to:
 ///   * TXN generation: requested_txns, created_txns, sign_failed_txns;
@@ -71,8 +51,6 @@ lazy_static! {
 pub struct Benchmarker {
     /// Using multiple clients can help improve the request speed.
     clients: Vec<Arc<AdmissionControlClient>>,
-    /// Use WalletLibrary to setup the benchmarking environment.
-    wallet: WalletLibrary,
     /// Interface to metric counters in validator nodes, e.g., #commited_txns in storage.
     debug_client: NodeDebugClient,
     /// Upper bound duration to stagger the clients before submitting TXNs.
@@ -92,12 +70,10 @@ impl Benchmarker {
         if clients.is_empty() {
             panic!("failed to create benchmarker without any AdmissionControlClient");
         }
-        let wallet = WalletLibrary::new();
         let arc_clients = clients.into_iter().map(Arc::new).collect();
         let prev_sequence_numbers = HashMap::new();
         Benchmarker {
             clients: arc_clients,
-            wallet,
             debug_client,
             stagger_range_ms,
             prev_sequence_numbers,
@@ -105,22 +81,8 @@ impl Benchmarker {
     }
 
     /// -------------------------------------------------------------------- ///
-    ///  Benchmark setup: account generation and minting APIs and helpers.   ///
+    ///  Benchmark setup: Load faucet account and minting APIs and helpers.  ///
     /// -------------------------------------------------------------------- ///
-
-    /// Create a new account without keypair from self's wallet.
-    fn gen_next_account(&mut self) -> AccountData {
-        let (address, _) = self
-            .wallet
-            .new_address()
-            .expect("failed to generate account address");
-        AccountData {
-            address,
-            key_pair: None,
-            sequence_number: 0,
-            status: AccountStatus::Local,
-        }
-    }
 
     /// Load keypair from given faucet_account_path,
     /// then try to sync with a validator to get up-to-date faucet account's sequence number.
@@ -149,28 +111,21 @@ impl Benchmarker {
         }
     }
 
-    /// Generate a number of random accounts and minting these accounts using self's AC client(s).
+    /// Minting given accounts using self's AC client(s).
     /// Mint TXNs must be 100% successful in order to continue benchmark.
-    /// Caller is responsible for terminating the benchmark with expect().
+    /// Therefore mint_accounts() will panic when any mint TXN is not accepted or fails.
     /// Known issue: Minting opereations from two different Benchmarker instances
     /// will fail because they are sharing the same faucet account.
-    pub fn gen_and_mint_accounts(
-        &mut self,
-        mint_key_file_path: &str,
-        num_accounts: u64,
-    ) -> Result<Vec<AccountData>> {
-        let mut results = vec![];
-        let mut mint_requests = vec![];
+    pub fn mint_accounts(&mut self, mint_key_file_path: &str, accounts: &[AccountData]) {
         let mut faucet_account = self.load_faucet_account(mint_key_file_path);
         self.prev_sequence_numbers
             .insert(faucet_account.address, faucet_account.sequence_number);
-        for _i in 0..num_accounts {
-            let account = self.gen_next_account();
-            let mint_txn_req = self.gen_mint_txn_request(&mut faucet_account, &account.address)?;
-            self.prev_sequence_numbers.insert(account.address, 0);
-            results.push(account);
-            mint_requests.push(mint_txn_req);
+        for account in accounts.iter() {
+            self.prev_sequence_numbers
+                .insert(account.address, account.sequence_number);
         }
+
+        let mint_requests = gen_mint_txn_requests(&mut faucet_account, accounts);
         // Disable client staggering for mint operations.
         let stagger_range_ms = self.stagger_range_ms;
         self.stagger_range_ms = 1;
@@ -179,125 +134,13 @@ impl Benchmarker {
         self.stagger_range_ms = stagger_range_ms;
         // We stop immediately if any minting fails.
         if num_accepted != mint_requests.len() || num_accepted - num_committed > 0 {
-            bail!(
+            panic!(
                 "{} of {} mint transaction(s) accepted, and {} failed",
                 num_accepted,
                 mint_requests.len(),
                 num_accepted - num_committed,
             )
-        } else {
-            Ok(results)
         }
-    }
-
-    /// ------------------------------------------ ///
-    ///  Transaction generation helpers and APIs.  ///
-    /// ------------------------------------------ ///
-
-    /// Craft a transaction request.
-    fn gen_submit_transaction_request(
-        &self,
-        program: Program,
-        sender_account: &mut AccountData,
-    ) -> Result<SubmitTransactionRequest> {
-        OP_COUNTER.inc("requested_txns");
-        let signer: Box<&dyn TransactionSigner> = match &sender_account.key_pair {
-            Some(key_pair) => Box::new(key_pair),
-            None => Box::new(&self.wallet),
-        };
-        // If generation fails here, sequence number will not be increased,
-        // so it is fine to continue later generation.
-        let signed_txn = create_signed_txn(
-            *signer,
-            program,
-            sender_account.address,
-            sender_account.sequence_number,
-            MAX_GAS_AMOUNT,
-            GAS_UNIT_PRICE,
-            TX_EXPIRATION,
-        )
-        .or_else(|e| {
-            OP_COUNTER.inc("sign_failed_txns");
-            Err(e)
-        })?;
-        let mut req = SubmitTransactionRequest::new();
-        req.set_signed_txn(signed_txn.into_proto());
-        sender_account.sequence_number += 1;
-        OP_COUNTER.inc("created_txns");
-        Ok(req)
-    }
-
-    /// Craft TXN request to mint receiver FREE_LUNCH libra coins.
-    fn gen_mint_txn_request(
-        &self,
-        faucet_account: &mut AccountData,
-        receiver: &AccountAddress,
-    ) -> Result<SubmitTransactionRequest> {
-        let program = vm_genesis::encode_mint_program(receiver, FREE_LUNCH);
-        self.gen_submit_transaction_request(program, faucet_account)
-    }
-
-    /// Craft TXN request to transfer coins from sender to receiver.
-    fn gen_transfer_txn_request(
-        &self,
-        sender: &mut AccountData,
-        receiver: &AccountAddress,
-        num_coins: u64,
-    ) -> Result<SubmitTransactionRequest> {
-        let program = vm_genesis::encode_transfer_program(&receiver, num_coins);
-        self.gen_submit_transaction_request(program, sender)
-    }
-
-    /// Pre-generate TXN requests of a circle of transfers.
-    /// For example, given account (A1, A2, A3, ..., AN), this method returns a vector of TXNs
-    /// like (A1->A2, A2->A3, A3->A4, ..., AN->A1).
-    pub fn gen_ring_txn_requests(
-        &self,
-        accounts: &mut [AccountData],
-    ) -> Vec<SubmitTransactionRequest> {
-        let mut receiver_addrs: Vec<AccountAddress> =
-            accounts.iter().map(|account| account.address).collect();
-        receiver_addrs.rotate_left(1);
-        accounts
-            .iter_mut()
-            .zip(receiver_addrs.iter())
-            .flat_map(|(sender, receiver_addr)| {
-                self.gen_transfer_txn_request(sender, receiver_addr, 1)
-                    .or_else(|e| {
-                        error!(
-                            "failed to generate {:?} to {:?} transfer TXN: {:?}",
-                            sender.address, receiver_addr, e
-                        );
-                        Err(e)
-                    })
-            })
-            .collect()
-    }
-
-    /// Pre-generate TXN requests of pairwise transfers between accounts, including self to self
-    /// transfer. For example, given account (A1, A2, A3, ..., AN), this method returns a vector
-    /// of TXNs like (A1->A1, A1->A2, ..., A1->AN, A2->A1, A2->A2, ... A2->AN, ..., AN->A(N-1)).
-    pub fn gen_pairwise_txn_requests(
-        &self,
-        accounts: &mut [AccountData],
-    ) -> Vec<SubmitTransactionRequest> {
-        let receiver_addrs: Vec<AccountAddress> =
-            accounts.iter().map(|account| account.address).collect();
-        let mut txn_reqs = vec![];
-        for sender in accounts.iter_mut() {
-            for receiver_addr in receiver_addrs.iter() {
-                match self.gen_transfer_txn_request(sender, receiver_addr, 1) {
-                    Ok(txn_req) => txn_reqs.push(txn_req),
-                    Err(e) => {
-                        error!(
-                            "failed to generate {:?} to {:?} transfer TXN: {:?}",
-                            sender.address, receiver_addr, e
-                        );
-                    }
-                }
-            }
-        }
-        txn_reqs
     }
 
     /// ----------------------------------------------------------------- ///
